quickly fix bad type before dalamud updates

+++ FFXIVClientStructs/FFXIV/Client/Game/InstanceContent/PublicContentDirector.cs
@@ -17,7 +17,7 @@ public unsafe partial struct PublicContentDirector {
     public static partial nint HandleEnterContentInfoPacket(EnterContentInfo* packet);
 }
 
-public enum PublicContentDirectorType : uint {
+public enum PublicContentDirectorType : byte {
     BondingCeremony = 1,
     TripleTriad = 2,
     Eureka = 3,
